# packages/shared/openapi.yaml
openapi: 3.0.0
paths:
...
  /api/v1/rfqs:
    post:
      operationId: RfqsController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRfqDto"
      responses:
        "201":
          description: ""
      security: &a9
        - bearer: []
      summary: Submit a new RFQ (multi-item)
      tags: &a10
        - rfqs
    get:
      operationId: RfqsController_list
      parameters: []
      responses:
        "200":
          description: ""
      security: *a9
      summary: List RFQs (own for buyers; all for staff/admin)
      tags: *a10
  /api/v1/rfqs/{id}:
    get:
      operationId: RfqsController_getOne
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
      security: *a9
      summary: RFQ detail with items, quotes and documents
      tags: *a10
  /api/v1/rfqs/{id}/transition:
    post:
      operationId: RfqsController_transition
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransitionRfqDto"
      responses:
        "201":
          description: ""
      security: *a9
      summary: Move RFQ through lifecycle (strict state machine; buyers may only
        cancel own RFQ)
      tags: *a10
  /api/v1/private-label:
    post:
      operationId: PrivateLabelController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePrivateLabelDto"
      responses:
        "201":
          description: ""
      security: &a11
        - bearer: []
      summary: Start a private-label/OEM request linked to my RFQ
      tags: &a12
        - private-label
  /api/v1/private-label/{id}:
    get:
      operationId: PrivateLabelController_getOne
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
      security: *a11
      summary: Private-label request detail
      tags: *a12
  /api/v1/private-label/{id}/transition:
    post:
      operationId: PrivateLabelController_transition
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransitionPlDto"
      responses:
        "201":
          description: ""
      security: *a11
      summary: Move private-label request forward (strict flow, staff/admin)
      tags: *a12
...
  /api/v1/quotes:
    post:
      operationId: QuotesController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateQuoteDto"
      responses:
        "201":
          description: ""
      security: &a15
        - bearer: []
      summary: Create a quote against an RFQ (staff/admin)
      tags: &a16
        - quotes
    get:
      operationId: QuotesController_list
      parameters: []
      responses:
        "200":
          description: ""
      security: *a15
      summary: List quotes (own for buyers; all for staff/admin)
      tags: *a16
  /api/v1/quotes/{id}/send:
    post:
      operationId: QuotesController_send
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "201":
          description: ""
      security: *a15
      summary: Send a draft quote to the buyer (staff/admin)
      tags: *a16
  /api/v1/quotes/{id}/accept:
    post:
      operationId: QuotesController_accept
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "201":
          description: ""
      security: *a15
      summary: Accept a live quote (buyer, owns the RFQ; creates an order)
      tags: *a16
  /api/v1/quotes/{id}/reject:
    post:
      operationId: QuotesController_reject
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RejectQuoteDto"
      responses:
        "201":
          description: ""
      security: *a15
      summary: Reject a live quote (buyer, owns the RFQ)
      tags: *a16
  /api/v1/quotes/{id}/request-revision:
    post:
      operationId: QuotesController_requestRevision
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RejectQuoteDto"
      responses:
        "201":
          description: ""
      security: *a15
      summary: Send a live quote back to staff for revision (buyer)
      tags: *a16
  /api/v1/quotes/{id}:
    get:
      operationId: QuotesController_getOne
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
      security: *a15
      summary: Quote detail
      tags: *a16
...
info:
  title: EGT B2B API
  description: Eagle Goods Trading Co. — sourcing & export backend API (v1)
  version: 1.0.0
  contact: {}
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
...
    CreateRfqItemDto:
      type: object
      properties:
        productId:
          type: string
          description: Catalogue product id (omit for custom "we source it" items)
        customName:
          type: string
        customDetails:
          type: string
        quantity:
          type: number
        unit:
          type: string
          default: units
      required:
        - quantity
    CreateRfqDto:
      type: object
      properties:
        contactName:
          type: string
        contactEmail:
          type: string
        contactPhone:
          type: string
        destinationCountry:
          type: string
        destinationCity:
          type: string
        deliveryAddress:
          type: string
        shippingPreference:
          type: string
          enum:
            - buyer_freight
            - egt_turnkey_freight
        currency:
          type: string
          default: USD
        budget:
          type: number
        targetPrice:
          type: number
        privateLabel:
          type: boolean
          default: false
        packagingPreference:
          type: string
        requirementDetails:
          type: string
          minLength: 15
        items:
          type: array
          items:
            $ref: "#/components/schemas/CreateRfqItemDto"
      required:
        - contactName
        - contactEmail
        - contactPhone
        - destinationCountry
        - requirementDetails
        - items
    TransitionRfqDto:
      type: object
      properties:
        to:
          type: string
          enum:
            - under_review
            - sourcing
            - supplier_matched
            - sample_discussion
            - quotation_ready
            - buyer_action_required
            - approved
            - order_processing
            - completed
            - closed
      required:
        - to
    CreatePrivateLabelDto:
      type: object
      properties:
        rfqId:
          type: string
        brandName:
          type: string
        designNotes:
          type: string
      required:
        - rfqId
    TransitionPlDto:
      type: object
      properties:
        to:
          type: string
          enum:
            - design_review
            - sample
            - approval
            - production
            - completed
      required:
        - to
...
    CreateQuoteItemDto:
      type: object
      properties:
        productId:
          type: string
        description:
          type: string
        quantity:
          type: number
        unit:
          type: string
          default: units
        unitPrice:
          type: number
      required:
        - description
        - quantity
        - unitPrice
    CreateQuoteDto:
      type: object
      properties:
        rfqId:
          type: string
        currency:
          type: string
          default: USD
        shippingCost:
          type: number
        incoterm:
          type: string
          enum:
            - FOB
            - CIF
            - DDP
            - EXW
        validUntil:
          type: string
          description: ISO date the quote remains valid until
        notes:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/CreateQuoteItemDto"
      required:
        - rfqId
        - validUntil
        - items
    RejectQuoteDto:
      type: object
      properties:
        reason:
          type: string
...